Factorize CommandAsset
