+++ www/sockets/src/Services/CommandAsset.php
@@ -555,109 +555,8 @@ class CommandAsset
         return false;
     }
     //GLOBAL USAGES FUNCTIONS -- END
-
-    //LS USAGES FUNCTIONS -- START
-    /**
-     * Get the files in the actual directory in an array
-     */
-    public static function getFiles(\PDO $db, string $terminal_mac, $currentPath)
-    {
-        $stmp = $db->prepare("SELECT name, chmod, editeddate, length(data), username, parent, idfile FROM terminal_file,terminal_user WHERE terminal_file.terminal=:mac AND parent=:parent AND idterminal_user = owner");
-        $stmp->bindParam(":mac", $terminal_mac);
-        $stmp->bindParam(":parent", $currentPath);
-        $stmp->execute();
-        $files = [];
-
-        while ($row = $stmp->fetch(\PDO::FETCH_ASSOC)) {
-            $files[] = Terminal_FileModel::map($row);
-        }
-        return $files;
-    }
-
-    public static function getDirectories(\PDO $db, string $terminal_mac, $currentPath)
-    {
-        $stmp = $db->prepare("SELECT name, chmod, editeddate, username, parent, iddir FROM terminal_directory,terminal_user WHERE terminal_directory.terminal=:mac AND parent=:parent AND idterminal_user = owner");
-        $stmp->bindParam(":mac", $terminal_mac);
-        $stmp->bindParam(":parent", $currentPath);
-        $stmp->execute();
-        $dirs = [];
-
-        while ($row = $stmp->fetch(\PDO::FETCH_ASSOC)) {
-            $dirs[] = Terminal_FileModel::map($row);
-        }
-        return $dirs;
-    }
-    //LS USAGES FUNCTIONS -- END
-
-    //RM USAGES FUNCTIONS -- START
-    public static function deleteFile(\PDO $db, SenderData &$data, ConnectionInterface $sender, string $terminal_mac, string $filename, int $parentId)
-    {
-        $stmp = $db->prepare("DELETE FROM terminal_file WHERE terminal = :terminal AND parent = :parent AND name = :name AND owner = :owner");
-
-        $stmp->bindParam(":terminal", $terminal_mac);
-        $stmp->bindParam(":parent", $parentId);
-        $stmp->bindParam(":name", $filename);
-        $stmp->bindParam(":owner", $data->user->idterminal_user);
-
-        return $stmp->execute();
-    }
-
-    public static function deleteDir(\PDO $db, SenderData &$data, ConnectionInterface $sender, string $terminal_mac, string $dirname, int $parentId)
-    {
-        $stmp = $db->prepare("DELETE FROM terminal_directory WHERE terminal = :terminal AND parent = :parent AND name = :name AND owner = :owner");
-
-        $stmp->bindParam(":terminal", $terminal_mac);
-        $stmp->bindParam(":parent", $parentId);
-        $stmp->bindParam(":name", $dirname);
-        $stmp->bindParam(":owner", $data->user->idterminal_user);
-
-        $stmp->execute();
-
-        $stmp = $db->prepare("SELECT name FROM terminal_directory WHERE terminal= :terminal AND parent= :parent AND name= :name AND owner= :owner");
-
-        //If the file or the dir exist, delete the file
-        $stmp->bindParam(":terminal", $terminal_mac);
-        $stmp->bindParam(":parent", $parentId);
-        $stmp->bindParam(":name", $dirname);
-        $stmp->bindParam(":owner", $data->user->idterminal_user);
-
-        $stmp->execute();
-        if ($stmp->fetch()['name']) {
-            $sender->send("message|<br> Directory not empty.");
-        };
-    }
-    //RM USAGES FUNCTIONS -- END
-
+    
     //TOUCH USAGES FUNCTIONS -- START
-    /**
-     * Full stage of creating new files
-     */
-    public static function stageCreateNewFiles(\PDO $db, SenderData &$data, ConnectionInterface $sender, string $terminal_mac, array $fullPathNewFiles)
-    {
-        foreach ($fullPathNewFiles as $fullPathNewFile) {
-            self::stageCreateNewFile($db, $data, $sender, $terminal_mac, $fullPathNewFile);
-        }
-    }
-
-    /**
-     * generate a new File
-     */
-    public static function createNewFile(\PDO $db, SenderData &$data, string $terminal_mac, string $name, int $parentId, string $content = ""): bool
-    {
-        $basicmod = 777;
-        $stmp = $db->prepare("INSERT INTO TERMINAL_FILE(terminal, parent, name, `data`, chmod, owner, `group`, createddate, editeddate) VALUES(:terminal, :parent, :name, :content, :chmod, :owner, (SELECT gid FROM terminal_user WHERE idterminal_user = :owner), NOW(),NOW());");
-
-        $stmp->bindParam(":terminal", $terminal_mac);
-        $stmp->bindParam(":parent", $parentId);
-        $stmp->bindParam(":name", $name);
-        $stmp->bindParam(":content", $content);
-        $stmp->bindParam(":chmod", $basicmod, \PDO::PARAM_INT);
-        $stmp->bindParam(":owner", $data->user->idterminal_user);
-
-        return
-        $stmp->execute();
-    }
-
     /**
      * Create or update files
      */
